+++ src/frontend/apps/impress/src/features/left-panel/components/LeftPanelCollapseButton.tsx
@@ -3,6 +3,7 @@ import { Button } from '@gouvfr-lasuite/cunningham-react';
 import { Text } from '@/components';
 import { FadeComponent } from '@/components/Effect';
 import { CardFloatingBar } from '@/components/FloatingBar';
+import { useResponsiveStore } from '@/stores';
 
 import LeftPanelIcon from '../assets/left-panel.svg';
 import { useLeftPanelStore } from '../stores';
@@ -15,6 +16,7 @@ export const LeftPanelCollapseButton = ({
   buttonTitle?: string;
 }) => {
   const { isPanelOpen, togglePanel } = useLeftPanelStore();
+  const { isSmallMobile } = useResponsiveStore();
 
   return (
     <CardFloatingBar className="--docs--left-panel-collapse-button">
@@ -27,17 +29,20 @@ export const LeftPanelCollapseButton = ({
         variant="tertiary"
         icon={<LeftPanelIcon width={24} height={24} aria-hidden="true" />}
         data-testid="floating-bar-toggle-left-panel"
-      ></Button>
-      <FadeComponent isVisible={!!buttonTitle}>
-        <Text
-          $size="sm"
-          $weight={700}
-          $color="var(--c--globals--colors--gray-1000)"
-          $padding={{ right: '2xs' }}
-        >
-          {buttonTitle}
-        </Text>
-      </FadeComponent>
+      >
+        {!isSmallMobile && !!buttonTitle && (
+          <FadeComponent isVisible={!!buttonTitle}>
+            <Text
+              $size="sm"
+              $weight={700}
+              $color="var(--c--globals--colors--gray-1000)"
+              title={buttonTitle}
+            >
+              {buttonTitle}
+            </Text>
+          </FadeComponent>
+        )}
+      </Button>
     </CardFloatingBar>
   );
 };